Add files via upload
Here I have a working demo

of the project. This project is called Guess My Number!

And I went with this kind of retro design

like inspired by the old 80s game.

Just to make it a bit more fun.

So as the name says,

the goal of this project

is to simply guess a secret number

which is between  1 and 20.